requirements.txt
python-telegram-bot==13.15
ultralytics
google-generativeai
Pillow
python-dotenv